# ai-stylist-starter/openapi.yaml
openapi: 3.1.0
info:
  title: AI Stylist API
  version: 0.1.0
servers:
  - url: http://localhost:8000
paths:
...
  /auth/oauth:
    post:
      summary: OAuth login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, token]
              properties:
                provider:
                  type: string
                  enum: [google, yandex, telegram]
                token:
                  type: string
      responses:
        "200":
          description: Auth success
  /auth/login:
    post:
      summary: Email login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Auth success
...
  /wardrobe/upload:
    post:
      summary: Upload wardrobe item
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image: { type: string, format: binary }
                category: { type: string }
      responses:
        "200":
          description: Item uploaded
...
  /wardrobe/confirm:
    post:
      summary: Confirm wardrobe item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [item_id, attributes]
              properties:
                item_id: { type: string }
                attributes: { type: object }
      responses:
        "200":
          description: Item updated
...